diesel compiles

+++ src/app.rs
@@ -84,7 +84,7 @@ impl App {
     }
 
     /// Deletes price history for the given Security, leaving only the latest price.
-    fn prune_for_sec(&self, security_id: i64) -> anyhow::Result<u16, Error> {
+    fn prune_for_sec(&self, security_id: i32) -> anyhow::Result<u16, Error> {
         // log::trace!("pruning prices for security id: {:?}", security_id);
 
         let mut count = 0;
